Resource manager properly prepped

+++ RTX_Dissertation/RTX_Dissertation/Mesh.cpp
@@ -4,13 +4,13 @@ Mesh::Mesh(ID3D12ResourcePtr vbo) : mVBO(vbo)
 	//Load BLAS?
 }
 
-int Mesh::AddInstance(mat4& transform)
+size_t Mesh::AddInstance(mat4& transform)
 {
 	mInstances.push_back(transform);
 	return mInstances.size() - 1;
 }
 
-void Mesh::RemoveInstance(int id)
+void Mesh::RemoveInstance(size_t id)
 {
 	if(id < mInstances.size())
 	{
